# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Colony App API
  description: A location-based social networking app API
  version: 1.0.0
  contact:
    name: Colony App Support
    email: support@colonyapp.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.colonyapp.com
    description: Production server
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                displayName:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      description: Authenticate user and get session tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      description: Sign out the current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logged out successfully
  /auth/user:
    get:
      tags: [Auth]
      summary: Get current user
      description: Get the authenticated user's details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /auth/profile:
    patch:
      tags: [Auth]
      summary: Update user profile
      description: Update display name and avatar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  example: New Name
                avatarUrl:
                  type: string
                  format: uri
                  example: https://example.com/avatar.jpg
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'

  # ==================== PROFILE ====================
  /profile/{userId}:
    get:
      tags: [Profile]
      summary: Get user profile by ID
      description: Fetch a user's profile by their ID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  profile:
                    $ref: '#/components/schemas/UserProfile'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile:
    patch:
      tags: [Profile]
      summary: Update my profile
      description: Update current user's profile (requires authentication)
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  example: "John Doe"
                bio:
                  type: string
                  example: "Hello, I'm John!"
                avatarUrl:
                  type: string
                  format: uri
                  example: "https://example.com/avatar.jpg"
                locationName:
                  type: string
                  example: "Vasant Colony, Rukunpura, Patna"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  profile:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== NEARBY USERS ====================
...
  /groups:
    get:
      tags: [Groups]
      summary: Get all groups
      description: Get all public groups with optional filters
      parameters:
      - name: latitude
        in: query
        schema:
          type: number
      - name: longitude
        in: query
        schema:
          type: number
      - name: radius
        in: query
        schema:
          type: number
          default: 5
      - name: limit
        in: query
        schema:
          type: integer
          default: 20
      - name: offset
        in: query
        schema:
          type: integer
          default: 0
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
    post:
      tags: [Groups]
      summary: Create group
      description: Create a new group (requires authentication)
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Tech Enthusiasts"
                description:
                  type: string
                  example: "A group for tech lovers"
                isPrivate:
                  type: boolean
                  default: false
                latitude:
                  type: number
                  example: 25.5941
                longitude:
                  type: number
                  example: 85.1376
                locationName:
                  type: string
                  example: "Patna, India"
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  group:
                    $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/{groupId}:
    get:
      tags: [Groups]
      summary: Get group by ID
      description: Get a specific group's details
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  group:
                    $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/{groupId}/join:
    post:
      tags: [Groups]
      summary: Join group
      description: Join a group as a member (requires authentication)
      security:
      - BearerAuth: []
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successfully joined group
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/{groupId}/leave:
    post:
      tags: [Groups]
      summary: Leave group
      description: Leave a group
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Left group
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        full_name:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        location_text:
          type: string
        latitude:
          type: number
        longitude:
          type: number
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        session:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresAt:
              type: integer
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string